refs #10365. Fix warnings.

+++ Code/Mantid/Framework/API/test/MatrixWorkspaceTest.h
@@ -730,13 +730,13 @@ public:
     // Signal is always 1 and volume of each box is 1. Therefore normalized signal values by volume should always be 1.
 
     // Test at the top right.
-    coord_t coord_top_right[2] = {ws.readX(0).back(),  double(0)};
+    coord_t coord_top_right[2] = {static_cast<float>(ws.readX(0).back()),  float(0)};
     signal_t value = 0;
     TS_ASSERT_THROWS_NOTHING(value = ws.getSignalAtCoord(coord_top_right, VolumeNormalization));
     TS_ASSERT_EQUALS(1.0, value);
 
     // Test at another location just to be sure.
-    coord_t coord_bottom_left[2] = {ws.readX(nVertical-1)[1],  double(nVertical-1) };
+    coord_t coord_bottom_left[2] = {static_cast<float>(ws.readX(nVertical-1)[1]),  float(nVertical-1) };
     TS_ASSERT_THROWS_NOTHING(value = ws.getSignalAtCoord(coord_bottom_left, VolumeNormalization));
     TS_ASSERT_EQUALS(1.0, value);
   }
